Extended system to support multiple actors
Extended system to support multiple actors. Please reference the foum post in the original commit to understand changes.
