+++ UOP1_Project/Assets/Art/Characters/PigChef/Animation/Character Expressions/Happy Stance.anim
@@ -313,7 +313,7 @@ AnimationClip:
     m_Level: 0
     m_CycleOffset: 0
     m_HasAdditiveReferencePose: 0
-    m_LoopTime: 0
+    m_LoopTime: 1
     m_LoopBlend: 0
     m_LoopBlendOrientation: 0
     m_LoopBlendPositionY: 0
