+++ UOP1_Project/Assets/Art/Materials/Character Expressions/Hamelt_Eye_Material.mat
@@ -32,7 +32,7 @@ Material:
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _MainTex:
-        m_Texture: {fileID: 2800000, guid: 7f5b6d7acebe2444a8f563014ea88ca2, type: 3}
+        m_Texture: {fileID: 2800000, guid: c91c29cbb5ba0e84db43e2983c6c44aa, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _MetallicGlossMap:
