+++ UOP1_Project/Assets/Localization Files/Asset Tables/Character Expresssions/Expression Dialogue Shared Data.asset
@@ -35,6 +35,22 @@ MonoBehaviour:
     m_Key: And win the game
     m_Metadata:
       m_Items: []
+  - m_Id: 963439475159040
+    m_Key: Poor little fellas
+    m_Metadata:
+      m_Items: []
+  - m_Id: 1066527776915456
+    m_Key: EVIL I'm Evil Hamlet
+    m_Metadata:
+      m_Items: []
+  - m_Id: 1069056606683136
+    m_Key: EVIL Good
+    m_Metadata:
+      m_Items: []
+  - m_Id: 1072094583013376
+    m_Key: EVIL Its Over
+    m_Metadata:
+      m_Items: []
   m_Metadata:
     m_Items: []
   m_KeyGenerator:
