+++ UOP1_Project/Assets/Localization Files/Asset Tables/Character Expresssions/Expression Dialogue_en.asset
@@ -27,15 +27,31 @@ MonoBehaviour:
     m_Metadata:
       m_Items: []
   - m_Id: 194842804224
-    m_Localized: I will destroy them with my cane!
+    m_Localized: I will destroy THEM with my cane!
     m_Metadata:
       m_Items: []
   - m_Id: 354276687872
-    m_Localized: And then I'll be free to cook amazing recipes!
+    m_Localized: But then I'll be free to cook amazing recipes!
     m_Metadata:
       m_Items: []
   - m_Id: 4162436284416
     m_Localized: And win the game!
+    m_Metadata:
+      m_Items: []
+  - m_Id: 963439475159040
+    m_Localized: Poor little fellas...
+    m_Metadata:
+      m_Items: []
+  - m_Id: 1066527776915456
+    m_Localized: I'm Evil Hamlet
+    m_Metadata:
+      m_Items: []
+  - m_Id: 1069056606683136
+    m_Localized: Good, I hope they destroy you!
+    m_Metadata:
+      m_Items: []
+  - m_Id: 1072094583013376
+    m_Localized: Well, then the fun is over for me ...
     m_Metadata:
       m_Items: []
   references:
