+++ UOP1_Project/Assets/Localization Files/Asset Tables/Character Expresssions/Phonemes Shared Data.asset
@@ -35,6 +35,22 @@ MonoBehaviour:
     m_Key: And win
     m_Metadata:
       m_Items: []
+  - m_Id: 963042907910144
+    m_Key: Poor little fellas
+    m_Metadata:
+      m_Items: []
+  - m_Id: 1066132111441920
+    m_Key: EVIL I'm Evil Hamlet
+    m_Metadata:
+      m_Items: []
+  - m_Id: 1068504518836224
+    m_Key: EVIL Good
+    m_Metadata:
+      m_Items: []
+  - m_Id: 1071556898406400
+    m_Key: EVIL Its Over
+    m_Metadata:
+      m_Items: []
   m_Metadata:
     m_Items: []
   m_KeyGenerator:
