+++ UOP1_Project/Assets/Localization Files/Asset Tables/Character Expresssions/Phonemes_en.asset
@@ -32,12 +32,28 @@ MonoBehaviour:
     m_Metadata:
       m_Items: []
   - m_Id: 2057495461888
-    m_Localized: EH . TH EH N . L L . B AH . F F R AH . T UW . K UH K . AH M Z NG
-      . R EH S AH P AH S S .
+    m_Localized: B UH . TH EH N . L L . B AH . F F R AH . T UW . K UH . AH M Z .
+      R EH S AH P AH S S .
     m_Metadata:
       m_Items: []
   - m_Id: 2482227462144
     m_Localized: EH N D . W IH N . TH AH . G EY M .
+    m_Metadata:
+      m_Items: []
+  - m_Id: 963042907910144
+    m_Localized: P P UH R . L IH L . F EH L L AH S S .
+    m_Metadata:
+      m_Items: []
+  - m_Id: 1066132111441920
+    m_Localized: AY M . IY V V L . HH M L EH T .
+    m_Metadata:
+      m_Items: []
+  - m_Id: 1068504518836224
+    m_Localized: G UH D . AY . HH OW P . TH AY . D UH S T R OY . Y OY .
+    m_Metadata:
+      m_Items: []
+  - m_Id: 1071556898406400
+    m_Localized: W L L . TH N . TH UH . F N . IH S . UH V R . F R . M IY .
     m_Metadata:
       m_Items: []
   references:
