+++ UOP1_Project/Assets/Localization Files/Asset Tables/Dialogue System/Actors/Actors Shared Data.asset
@@ -27,6 +27,10 @@ MonoBehaviour:
     m_Key: PigChef
     m_Metadata:
       m_Items: []
+  - m_Id: 26988952259477504
+    m_Key: EvilPigChef
+    m_Metadata:
+      m_Items: []
   m_Metadata:
     m_Items: []
   m_KeyGenerator:
