+++ UOP1_Project/Assets/Localization Files/Asset Tables/Dialogue System/Actors/Actors_fr.asset
@@ -30,5 +30,9 @@ MonoBehaviour:
     m_Localized: Hamlet
     m_Metadata:
       m_Items: []
+  - m_Id: 26988952259477504
+    m_Localized: Evil Hamlet
+    m_Metadata:
+      m_Items: []
   references:
     version: 1
