+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Dialogue Line 004.asset
@@ -17,11 +17,11 @@ MonoBehaviour:
     m_TableReference:
       m_TableCollectionName: GUID:c89dcd31aa56e05408dec38505e1571d
     m_TableEntryReference:
-      m_KeyId: 354276687872
+      m_KeyId: 963439475159040
       m_Key: 
   _phonemeSentence:
     m_TableReference:
       m_TableCollectionName: GUID:5b34c1d30b339424984a3e4f37a292f0
     m_TableEntryReference:
-      m_KeyId: 2057495461888
+      m_KeyId: 963042907910144
       m_Key:
