+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Dialogue Line 005.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 52bc34e6fec986f4593b37da5d8be6d8
+guid: 404dedc930f497a43bb9a3b37df46901
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
