+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Dialogue Line 006.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 95a2c9434fe137544863393349fb5727
+guid: 52bc34e6fec986f4593b37da5d8be6d8
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
