+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Evil Dialogue Line 001.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 95a2c9434fe137544863393349fb5727
+guid: 1d4e6c8e7b79d3c49a50e626ab141cd7
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
