+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Evil Dialogue Line 002.asset
@@ -10,18 +10,18 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: e49a73eb0802e10429213abfb59e46e1, type: 3}
-  m_Name: Dialogue Line 005
+  m_Name: Evil Dialogue Line 002
   m_EditorClassIdentifier: 
-  _actor: {fileID: 11400000, guid: c7902af5c8e25244f94914ce51e7ea8a, type: 2}
+  _actor: {fileID: 11400000, guid: 2864cd188da1bc445875ce116fa264f6, type: 2}
   _sentence:
     m_TableReference:
       m_TableCollectionName: GUID:c89dcd31aa56e05408dec38505e1571d
     m_TableEntryReference:
-      m_KeyId: 354276687872
+      m_KeyId: 1069056606683136
       m_Key: 
   _phonemeSentence:
     m_TableReference:
       m_TableCollectionName: GUID:5b34c1d30b339424984a3e4f37a292f0
     m_TableEntryReference:
-      m_KeyId: 2057495461888
+      m_KeyId: 1068504518836224
       m_Key:
