+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Evil Dialogue Line 002.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 95a2c9434fe137544863393349fb5727
+guid: bd147aedf9260bb438c19b4ae0a36a0f
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
