+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Evil Dialogue Line 003.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 95a2c9434fe137544863393349fb5727
+guid: f37c15c5298f9884eae76c722b345710
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
