+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Mood Collections/[Evil Hamlet - Angry] Mood Collection.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 95a2c9434fe137544863393349fb5727
+guid: e74150c28be5bc842b5b73d12dc3695d
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
