+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Mood Collections/[Evil Hamlet - Explaining] Mood Collection.asset
@@ -1,0 +1,27 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 06b420d86736034438f480afde7aad1a, type: 3}
+  m_Name: '[Evil Hamlet - Explaining] Mood Collection'
+  m_EditorClassIdentifier: 
+  Actor: {fileID: 11400000, guid: 2864cd188da1bc445875ce116fa264f6, type: 2}
+  Mood: 7
+  EyeState_Open: {fileID: 2800000, guid: 576e41fe26f1c6c4fb4e02eb05552c4f, type: 3}
+  EyeState_MidBlink: {fileID: 2800000, guid: 72d1a733711fdc2448a4528010ab3785, type: 3}
+  EyeState_Closed: {fileID: 2800000, guid: 4968ec7dfb5b5174597d5ae74db7d6b9, type: 3}
+  LocalizedPhonemeSet:
+    m_TableReference:
+      m_TableCollectionName: GUID:99f15d4b14082d345b7ae040654c0b2a
+    m_TableEntryReference:
+      m_KeyId: 71916142592
+      m_Key: 
+  AnimatorClipTitles:
+  - Fist Pump
