+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Mood Collections/[Evil Hamlet - Explaining] Mood Collection.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 95a2c9434fe137544863393349fb5727
+guid: f89d6add4211a7643b7f0e283aaa5b33
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
