+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Mood Collections/[Evil Hamlet - Happy] Mood Collection.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 95a2c9434fe137544863393349fb5727
+guid: 1285306183583eb40907d763e9c27cd6
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
