+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Mood Collections/[Evil Hamlet - Neutral] Mood Collection.asset
@@ -1,0 +1,26 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 06b420d86736034438f480afde7aad1a, type: 3}
+  m_Name: '[Evil Hamlet - Neutral] Mood Collection'
+  m_EditorClassIdentifier: 
+  Actor: {fileID: 11400000, guid: 2864cd188da1bc445875ce116fa264f6, type: 2}
+  Mood: 1
+  EyeState_Open: {fileID: 2800000, guid: f4ac4be28316975438e1fc7b785f0453, type: 3}
+  EyeState_MidBlink: {fileID: 2800000, guid: a0546ff38360a974392655a9f7802115, type: 3}
+  EyeState_Closed: {fileID: 2800000, guid: 4440ee0823ba0c74aa1f980cdb583fc9, type: 3}
+  LocalizedPhonemeSet:
+    m_TableReference:
+      m_TableCollectionName: GUID:99f15d4b14082d345b7ae040654c0b2a
+    m_TableEntryReference:
+      m_KeyId: 71916142592
+      m_Key: 
+  AnimatorClipTitles: []
