+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Mood Collections/[Evil Hamlet - Neutral] Mood Collection.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 95a2c9434fe137544863393349fb5727
+guid: b5ff19b6c9e6fb049b0b10669f471078
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
