+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Mood Collections/[Evil Hamlet - Sad] Mood Collection.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 95a2c9434fe137544863393349fb5727
+guid: 7448256c4dec35742b08a699d5eb21ba
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
