+++ UOP1_Project/Assets/ScriptableObjects/Narrative/Character Expression Data/Mood Collections/[Evil Hamlet - Scared] Mood Collection.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 95a2c9434fe137544863393349fb5727
+guid: c67b5d0015140ae41a7a1eabcffef02a
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
