+++ UOP1_Project/Assets/Scripts/Characters/Character Expressions/ExpressionManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ba05cd28e7c958b43abafc49ab28fbdc
+guid: 43951d5d6c43e5049840abd4ca47e81e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
