--- UOP1_Project/Assets/Scripts/Characters/Character Expressions/ExpressionManagerMulti.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 43951d5d6c43e5049840abd4ca47e81e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
