+++ UOP1_Project/Assets/Scripts/Cutscenes/CharacterMoodControlTrack/CharacterMoodTrack.cs
@@ -5,7 +5,6 @@ using UnityEngine.Timeline;
 using UnityEngine.Playables;
 
 [TrackBindingType(typeof(ExpressionManager))]
-//[TrackBindingType(typeof(ExpressionManagerMulti))]
 [TrackClipType(typeof(CharacterMoodClip))]
 [TrackColor(0.0f, 0.0f, 0.0f)]
 public class CharacterMoodTrack : PlayableTrack
@@ -13,13 +12,11 @@ public class CharacterMoodTrack : PlayableTrack
 	public override Playable CreateTrackMixer(PlayableGraph graph, GameObject go, int inputCount)
 	{
 		ExpressionManager em = go.GetComponent<PlayableDirector>().GetGenericBinding(this) as ExpressionManager;
-		//ExpressionManagerMulti em = go.GetComponent<PlayableDirector>().GetGenericBinding(this) as ExpressionManagerMulti;
 
 		foreach (TimelineClip clip in GetClips())
 		{
 			CharacterMoodClip moodClip = clip.asset as CharacterMoodClip;
 
-			//moodClip.ExpressionManager = em;
 			moodClip.ExpressionManager = em;
 		}
