+++ UOP1_Project/Assets/Scripts/Cutscenes/CutsceneManager.cs
@@ -7,12 +7,10 @@ public class CutsceneManager : MonoBehaviour
 	[SerializeField] private InputReader _inputReader = default;
 	[SerializeField] private DialogueManager _dialogueManager = default;
 	[SerializeField] private ExpressionManager _expressionManager = default;
-	//[SerializeField] private ExpressionManagerMulti _expressionManager = default;
 
 	private PlayableDirector _activePlayableDirector;
 	private bool _isPaused;
 
-	//public ExpressionManagerMulti ExpressionManager { get => _expressionManager; set => _expressionManager = value; }
 	public ExpressionManager ExpressionManager { get => _expressionManager; set => _expressionManager = value; }
 	public bool IsCutscenePlaying => _activePlayableDirector.playableGraph.GetRootPlayable(0).GetSpeed() != 0d;
